backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.18
PyMuPDF>=1.25.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
